+++ Assets/Scripts/Interactables/Potions/HealthPotion.cs
@@ -4,15 +4,15 @@ using UnityEngine;
 
 public class HealthPotion : MonoBehaviour
 {
-    [SerializeField] private float healthRestore = 10;
+    [SerializeField] private int healthRestore = 10;
 
     // Check to see if the player is next to the potion and is pressing the interact key
     private void OnTriggerStay2D(Collider2D collision)
     {
         if (collision.tag == "Player" && Input.GetKeyDown(KeyCode.F))
         {
-            collision.GetComponent<PlayerState>().IncreaseHealth(healthRestore);            // Restore the player's health by a certain amount      
-            Destroy(this.gameObject);
+            collision.GetComponent<PlayerState>().IncreaseHealth(healthRestore);            // Restore the player's health by a specified amount
+            Destroy(this.gameObject);                                                       // Remove the potion from the level
         }
     }
 }
